+++ db/migrate/20260924100000_add_max_responses_to_surveys.rb
@@ -1,0 +1,5 @@
+class AddMaxResponsesToSurveys < ActiveRecord::Migration[8.1]
+  def change
+    add_column :surveys, :max_responses, :integer
+  end
+end